@@ -45,9 +45,9 @@ else
   end
 end
 
-gem_package "passenger" do
-  version node[:passenger][:version]
-end
+# gem_package "passenger" do
+#   version node[:passenger][:version]
+# end
 
 execute "passenger_module" do
   command 'passenger-install-apache2-module --auto'
