@@ -2,83 +2,105 @@ require_recipe "apt"
 require_recipe "build-essential"
 require_recipe "git"
 
-require_recipe "apache2"
-require_recipe "passenger_apache2::mod_rails"
+# Install rvm and Ruby 1.9.3. Add Chef Solo path
+require_recipe "rvm::system"
+require_recipe "rvm::vagrant"
 
 require 'securerandom'
 require 'yaml'
 
 ::Chef::Recipe.send(:include, Opscode::OpenSSL::Password)
 
-# Install more recent version of chef gem
-gem_package "chef" do
-  version "0.10.10"
-  action :install
+# Generate new password for MySQL root unless it has already been stored in database.yml 
+# This has to go before the require_recipe for mysql::server
+if File.exists? "/vagrant/config/database.yml"
+  stored_password = YAML.load_file("/vagrant/config/database.yml")["test"]["password"]
+  node.set['mysql']['server_root_password'] = stored_password
+else
+  # create new database.yml
+  node.set['mysql']['server_root_password'] = secure_password 
+  template "/vagrant/config/database.yml" do
+    source 'database.yml.erb'
+    owner 'root'
+    group 'root'
+    mode 0644
+  end
 end
 
-# Install bundler gem and run bundle command
-gem_package "bundler" do
-  version "1.1.3"
-  action :install
+require_recipe "mysql::server"
+
+# Install CouchDB and create default CouchDB database
+require_recipe "couchdb"
+execute "create CouchDB database #{node[:couchdb][:db_name]}" do
+  command "curl -X PUT http://127.0.0.1:5984/#{node[:couchdb][:db_name]}/"
+  ignore_failure true
 end
+
+
+
+# Generate new keys unless they have already been stored in settings.yml
+if File.exists? "/vagrant/config/settings.yml"
+  settings = YAML.load_file("/vagrant/config/settings.yml")["defaults"]
+  node.set_unless['app']['key'] = settings["rest_auth_site_key"]
+  node.set_unless['app']['secret'] = settings["session_secret"]
+else
+  # create new settings.yml
+  node.set_unless['app']['key'] = SecureRandom.hex(30)
+  node.set_unless['app']['secret'] = SecureRandom.hex(30)
+  template "/vagrant/config/settings.yml" do
+    source 'settings.yml.erb'
+    owner 'root'
+    group 'root'
+    mode 0644
+  end
+end
+
+# create new source_configs.yml
+unless File.exists? "/vagrant/config/source_configs.yml"
+  template "/vagrant/config/source_configs.yml" do
+    source 'source_configs.yml.erb'
+    owner 'root'
+    group 'root'
+    mode 0644
+  end
+end
+
+# Run bundle command
 bash "run bundle install in app directory" do
   cwd "/vagrant"
   code "bundle install"
 end
 
-# Generate new password for MySQL root unless it has already been stored in database.yml 
-# This has to go before the require_recipe for mysql::server
-if File.exists? "/vagrant/config/database.yml"
-  stored_password = YAML.load_file("/vagrant/config/database.yml")["test"]["password"]
-  node.set_unless['mysql']['server_root_password'] = stored_password
-else
-  node.set_unless['mysql']['server_root_password'] = secure_password  
-end
-
-require_recipe "mysql::server"
-
-# create new database.yml
-template "/vagrant/config/database.yml" do
-  source 'database.yml.erb'
+# Optionally seed the database with sources, groups and sample articles
+template "/vagrant/db/seeds.rb" do
+  source 'seeds.rb.erb'
   owner 'root'
   group 'root'
   mode 0644
 end
 
-# Generate new password for MySQL root unless it has already been stored in database.yml 
-# This has to go before the require_recipe for mysql::server
-if File.exists? "/vagrant/config/settings.yml"
-  settings = YAML.load_file("/vagrant/config/database.yml")["defaults"]
-  node.set_unless['app']['key'] = settings["rest_auth_site_key"]
-  node.set_unless['app']['secret'] = settings["session_secret"]
-else
-  node.set_unless['app']['key'] = SecureRandom.hex(30)
-  node.set_unless['app']['secret'] = SecureRandom.hex(30)
-  
-end
-
-# create new settings.yml
-template "/vagrant/config/settings.yml" do
-
-  source 'settings.yml.erb'
-  owner 'root'
-  group 'root'
-  mode 0644
-end
-
-# Create default database and run migrations
-bash "RAILS_ENV=#{node[:rails][:environment]} rake db:setup" do
+# Create default databases and run migrations
+bash "rake db:setup RAILS_ENV=#{node[:rails][:environment]}" do
   cwd "/vagrant"
-  code "RAILS_ENV=#{node[:rails][:environment]} rake db:setup"
+  code "rake db:setup RAILS_ENV=#{node[:rails][:environment]}"
 end
+
+# Start workers
+bash "rake workers:start_all RAILS_ENV=#{node[:rails][:environment]}" do
+  cwd "/vagrant"
+  code "rake workers:start_all RAILS_ENV=#{node[:rails][:environment]}"
+end
+
+# Install Apache and Passenger
+require_recipe "passenger_apache2::mod_rails"
 
 execute "disable-default-site" do
   command "sudo a2dissite default"
   notifies :reload, resources(:service => "apache2"), :delayed
 end
 
-web_app "default" do
+web_app "alm" do
   docroot "/vagrant/public"
-  template "default.erb"
+  template "alm.conf.erb"
   notifies :reload, resources(:service => "apache2"), :delayed
 end
